requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
azure-kusto-data==4.5.1
truststore==0.10.4

langchain
langchain-community
langchain-ollama